move `RowsFrameBounds` and `RangeFrameBounds` to separated modules
