+++ src/expr/core/src/window_function/mod.rs
@@ -17,6 +17,10 @@ pub use kind::*;
 
 mod call;
 pub use call::*;
+mod rows;
+pub use rows::*;
+mod range;
+pub use range::*;
 
 mod state;
 pub use state::*;
